update: rebuild keyed dioxus with build optimizations
