+++ frameworks/keyed/dioxus/index.html
@@ -2,13 +2,14 @@
     <meta charset="UTF-8">
     <title>Dioxus</title>
     <link href="/css/currentStyle.css" rel="stylesheet">
+    <script type="module">import init from '/frameworks/keyed/dioxus/bundled-dist/dioxus_benchmark-2AF7289E-E58F-4668-9666-1072F163C8AA.js';init('/frameworks/keyed/dioxus/bundled-dist/dioxus_benchmark-2AF7289E-E58F-4668-9666-1072F163C8AA_bg.wasm');</script>    
 
-<link rel="preload" href="/frameworks/non-keyed/dioxus/bundled-dist/dioxus_benchmark-E89064E7-2331-4235-94DD-EF3F5D4A839D_bg.wasm" as="fetch" type="application/wasm" crossorigin="">
-<link rel="modulepreload" href="/frameworks/non-keyed/dioxus/bundled-dist/dioxus_benchmark-E89064E7-2331-4235-94DD-EF3F5D4A839D.js"></head>
+<link rel="preload" href="/frameworks/keyed/dioxus/bundled-dist/dioxus_benchmark-2AF7289E-E58F-4668-9666-1072F163C8AA_bg.wasm" as="fetch" type="application/wasm" crossorigin="">
+<link rel="modulepreload" href="/frameworks/keyed/dioxus/bundled-dist/dioxus_benchmark-2AF7289E-E58F-4668-9666-1072F163C8AA.js"></head>
 <body>
 <div id="main"></div>
 
-<script src="/frameworks/non-keyed/dioxus/isKeyed.js"></script>
+<script src="/frameworks/keyed/dioxus/isKeyed.js"></script>
 
 
-<script type="module">import init from '/frameworks/non-keyed/dioxus/bundled-dist/dioxus_benchmark-E89064E7-2331-4235-94DD-EF3F5D4A839D.js';init('/frameworks/non-keyed/dioxus/bundled-dist/dioxus_benchmark-E89064E7-2331-4235-94DD-EF3F5D4A839D_bg.wasm');</script></body></html>
+</body></html>
