+++ frameworks/keyed/dioxus/trunk.html
@@ -9,6 +9,6 @@
 <body>
 <div id='main'></div>
 
-<script src="/frameworks/non-keyed/dioxus/isKeyed.js"></script>
+<script src="/frameworks/keyed/dioxus/isKeyed.js"></script>
 </body>
 </html>
